Adds a user service to app: sign/log in, auth, add/remove tokens
The userService checks if a token is still good the first time it's
instantiated -- if the token is old, it's removed.

The return object for the service has a method to check if a user
is authenticated, to log out (remove their token), and a dual-purpose
method to handle log-in and sign-up requests (they use the same data
but different end-points on the API).

Follows Marty's format closely.