Link to new store in add now works; Adds spacing styling to form
The $state.go wasn't going to the new store, despite the fact the id
was available and loggable. It didn't seem to be saving such that
I could use it outside of the promise chain. Instead, moved the
reset and link to inside the then() method and now the add works!

The styling is to have the form not be all on one line. Looks
better.